requirements.txt
# Django and core dependencies
Django==4.2.11
python-decouple==3.8
whitenoise==6.6.0
gunicorn==21.2.0
Pillow
dj-database-url==2.1.0
mysqlclient==2.2.4

# Email and forms
django-crispy-forms==2.1
crispy-bootstrap5==2024.2

# Development tools
django-debug-toolbar==4.2.0
django-compressor==4.5.1
django-libsass==0.9
libsass==0.22.0

# Security and utilities
django-cors-headers==4.3.1
django-extensions==3.2.3
requests==2.31.0

# SEO and analytics
django-meta==2.4.0

# File handling and media (optional for production)
# boto3==1.34.34
# django-storages==1.14.2
